+++ CHANGES.rst
@@ -10,6 +10,12 @@ Released 2025-02-22
 
 
 
+Unreleased
+----------
+
+- Update ``intersphinx_mapping`` for Sphinx 8 compatibility. :pr:`599`, :issue:`598`
+
+
 Version 2.3.0
 -------------