+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changelog
 Version 2.4.0
 -------------
 
-Unreleased
+2026-04-17
 
 - Modernize project setup
 - Update ``intersphinx_mapping`` for Sphinx 8 compatibility. :pr:`599`, :issue:`598`